Combine Account and Sync settings sections; move user management out
Account and Sync now render under a single "Account" section since
they're both about the signed-in user's account state, cutting a
redundant nav entry. The admin-only user management table moves out
of Account into its own gated "User management" section, which only
appears in the Settings nav for admins.

Closes #1001
