+++ frontend/src/components/settings/account/SettingsAccountSection.tsx
@@ -4,22 +4,8 @@ import Alert from "react-bootstrap/Alert";
 import Button from "react-bootstrap/Button";
 import Form from "react-bootstrap/Form";
 import ListGroup from "react-bootstrap/ListGroup";
-import Table from "react-bootstrap/Table";
 import { ConfirmationDialog } from "@/components/ConfirmationDialog";
 import * as m from "@/paraglide/messages.js";
-import { getLocale } from "@/paraglide/runtime.js";
-
-const formatTimestamp = (value: string): string => {
-  const parsed = new Date(value);
-  if (Number.isNaN(parsed.getTime())) {
-    return value;
-  }
-  const locale = getLocale() === "nl" ? "nl-NL" : "en-US";
-  return new Intl.DateTimeFormat(locale, {
-    dateStyle: "medium",
-    timeStyle: "short",
-  }).format(parsed);
-};
 
 interface SettingsAccountSectionProps {
   isValidating: boolean;
@@ -34,22 +20,10 @@ interface SettingsAccountSectionProps {
   profileDraft: string;
   isProfileSaving: boolean;
   hasProfileChanges: boolean;
-  adminUsers: Array<{
-    id: number;
-    username: string;
-    display_name: string;
-    created_at: string;
-    updated_at: string;
-  }>;
-  isAdminUsersLoading: boolean;
-  adminUsersError: string | null;
-  adminUsersDeleteError: string | null;
-  deletingAdminUserId: number | null;
   isDeletingAccount: boolean;
   deleteAccountError: string | null;
   onProfileDraftChange: (value: string) => void;
   onSaveProfile: () => void;
-  onDeleteAdminUser: (userId: number) => void;
   onDeleteAccount: () => void;
   onLogout: () => void;
   onSignup: () => void;
@@ -69,23 +43,15 @@ export function SettingsAccountSection({
   profileDraft,
   isProfileSaving,
   hasProfileChanges,
-  adminUsers,
-  isAdminUsersLoading,
-  adminUsersError,
-  adminUsersDeleteError,
-  deletingAdminUserId,
   isDeletingAccount,
   deleteAccountError,
   onProfileDraftChange,
   onSaveProfile,
-  onDeleteAdminUser,
   onDeleteAccount,
   onLogout,
   onSignup,
   onLogin,
 }: SettingsAccountSectionProps) {
-  const [pendingDeleteUserId, setPendingDeleteUserId] = useState<number | null>(null);
-  const pendingDeleteUser = adminUsers.find((user) => user.id === pendingDeleteUserId) ?? null;
   const [showDeleteAccountConfirm, setShowDeleteAccountConfirm] = useState(false);
 
   return (
@@ -187,81 +153,6 @@ export function SettingsAccountSection({
                       </Button>
                     </div>
 
-                    {isAdmin ? (
-                      <div className="pt-2 border-top">
-                        <h6 className="mb-2">{m.account_admin_users_title()}</h6>
-                        <p className="text-muted small mb-2">{m.account_admin_users_description()}</p>
-                        {isAdminUsersLoading ? (
-                          <div className="d-flex align-items-center gap-2 text-muted small">
-                            <span
-                              className="spinner-border spinner-border-sm"
-                              role="status"
-                              aria-hidden="true"
-                            ></span>
-                            <span>{m.account_admin_users_loading()}</span>
-                          </div>
-                        ) : adminUsersError ? (
-                          <Alert variant="warning" className="mb-0 py-2">
-                            {adminUsersError}
-                          </Alert>
-                        ) : (
-                          <>
-                            {adminUsersDeleteError ? (
-                              <Alert variant="danger" className="mb-2 py-2">
-                                {adminUsersDeleteError}
-                              </Alert>
-                            ) : null}
-                            {adminUsers.length === 0 ? (
-                              <p className="text-muted small mb-0">{m.account_admin_users_empty()}</p>
-                            ) : (
-                              <div className="table-responsive">
-                                <Table size="sm" striped hover className="mb-0 align-middle">
-                                  <thead>
-                                    <tr>
-                                      <th>{m.account_admin_users_user_id()}</th>
-                                      <th>{m.account_admin_users_username()}</th>
-                                      <th>{m.account_admin_users_display_name()}</th>
-                                      <th>{m.account_admin_users_created_at()}</th>
-                                      <th>{m.account_admin_users_updated_at()}</th>
-                                      <th className="text-end">{m.account_admin_users_actions()}</th>
-                                    </tr>
-                                  </thead>
-                                  <tbody>
-                                    {adminUsers.map((user) => (
-                                      <tr key={user.id}>
-                                        <td>{user.id}</td>
-                                        <td>{user.username}</td>
-                                        <td>{user.display_name}</td>
-                                        <td>{formatTimestamp(user.created_at)}</td>
-                                        <td>{formatTimestamp(user.updated_at)}</td>
-                                        <td className="text-end">
-                                          <Button
-                                            variant="outline-danger"
-                                            size="sm"
-                                            disabled={accountId === user.id || deletingAdminUserId !== null}
-                                            title={
-                                              accountId === user.id
-                                                ? m.account_admin_users_delete_self_blocked()
-                                                : undefined
-                                            }
-                                            onClick={() => setPendingDeleteUserId(user.id)}
-                                          >
-                                            {deletingAdminUserId === user.id
-                                              ? m.account_admin_users_delete_busy()
-                                              : m.delete()}
-                                          </Button>
-                                        </td>
-                                      </tr>
-                                    ))}
-                                  </tbody>
-                                </Table>
-                              </div>
-                            )}
-                          </>
-                        )}
-                      </div>
-                    ) : null}
-
                     <div className="pt-2 border-top">
                       <h6 className="mb-2 text-danger">{m.account_delete_section_title()}</h6>
                       <p className="text-muted small mb-2">{m.account_delete_description()}</p>
@@ -317,30 +208,6 @@ export function SettingsAccountSection({
           )}
         </ListGroup>
       </div>
-      <ConfirmationDialog
-        isOpen={pendingDeleteUser !== null}
-        title={m.account_admin_users_delete_confirm_title()}
-        message={
-          pendingDeleteUser
-            ? m.account_admin_users_delete_confirm_message({
-                name: pendingDeleteUser.display_name || pendingDeleteUser.username,
-                username: pendingDeleteUser.username,
-              })
-            : ""
-        }
-        confirmLabel={m.delete()}
-        cancelLabel={m.cancel()}
-        onConfirm={() => {
-          if (!pendingDeleteUser) {
-            return;
-          }
-          onDeleteAdminUser(pendingDeleteUser.id);
-          setPendingDeleteUserId(null);
-        }}
-        onCancel={() => setPendingDeleteUserId(null)}
-        variant="danger"
-        icon="bi-trash"
-      />
       <ConfirmationDialog
         isOpen={showDeleteAccountConfirm}
         title={m.account_delete_confirm_title()}
