+++ frontend/src/pages/settings/hooks/useSettingsAccount.ts
@@ -1,6 +1,7 @@
 import { useEffect, useState } from "react";
 import * as m from "@/paraglide/messages.js";
 import { logger } from "@/utils/logger";
+import { readErrorDetail } from "@/utils/apiClient";
 
 interface AccountProfile {
   id: number;
@@ -12,14 +13,6 @@ interface AccountProfile {
   };
 }
 
-interface ManagedUser {
-  id: number;
-  username: string;
-  display_name: string;
-  created_at: string;
-  updated_at: string;
-}
-
 interface UseSettingsAccountParams {
   isAuthenticated: boolean;
   displayName: string | null;
@@ -27,17 +20,6 @@ interface UseSettingsAccountParams {
   showSuccessToast: (message: string, icon?: string) => void;
   onAccountDeleted: () => void;
 }
-
-const readErrorDetail = async (response: Response): Promise<string | null> => {
-  try {
-    const payload = (await response.json()) as { detail?: unknown };
-    return typeof payload.detail === "string" && payload.detail.trim() !== ""
-      ? payload.detail
-      : null;
-  } catch {
-    return null;
-  }
-};
 
 export function useSettingsAccount({
   isAuthenticated,
@@ -51,11 +33,6 @@ export function useSettingsAccount({
   const [isProfileLoading, setIsProfileLoading] = useState(false);
   const [isProfileSaving, setIsProfileSaving] = useState(false);
   const [profileError, setProfileError] = useState<string | null>(null);
-  const [adminUsers, setAdminUsers] = useState<ManagedUser[]>([]);
-  const [isAdminUsersLoading, setIsAdminUsersLoading] = useState(false);
-  const [adminUsersError, setAdminUsersError] = useState<string | null>(null);
-  const [adminUsersDeleteError, setAdminUsersDeleteError] = useState<string | null>(null);
-  const [deletingAdminUserId, setDeletingAdminUserId] = useState<number | null>(null);
   const [isDeletingAccount, setIsDeletingAccount] = useState(false);
   const [deleteAccountError, setDeleteAccountError] = useState<string | null>(null);
 
@@ -97,48 +74,6 @@ export function useSettingsAccount({
       isCancelled = true;
     };
   }, [isAuthenticated, fetchFn]);
-
-  useEffect(() => {
-    if (!isAuthenticated || !accountProfile?.is_admin) {
-      setAdminUsers([]);
-      setAdminUsersError(null);
-      setAdminUsersDeleteError(null);
-      setIsAdminUsersLoading(false);
-      setDeletingAdminUserId(null);
-      return;
-    }
-
-    let isCancelled = false;
-    setIsAdminUsersLoading(true);
-    setAdminUsersError(null);
-    setAdminUsersDeleteError(null);
-
-    fetchFn("/api/users/?limit=100")
-      .then(async (response) => {
-        if (!response.ok) {
-          throw new Error(`Unexpected status: ${response.status}`);
-        }
-        const payload = (await response.json()) as {
-          items?: ManagedUser[];
-        };
-        if (isCancelled) return;
-        setAdminUsers(payload.items ?? []);
-      })
-      .catch((error) => {
-        if (isCancelled) return;
-        logger.error("Failed to load admin user list:", error);
-        setAdminUsersError(m.account_admin_users_load_failed());
-      })
-      .finally(() => {
-        if (!isCancelled) {
-          setIsAdminUsersLoading(false);
-        }
-      });
-
-    return () => {
-      isCancelled = true;
-    };
-  }, [isAuthenticated, accountProfile?.is_admin, fetchFn]);
 
   const hasProfileChanges =
     accountProfile !== null &&
@@ -191,37 +126,6 @@ export function useSettingsAccount({
     }
   };
 
-  const handleDeleteAdminUser = async (userId: number) => {
-    if (accountProfile?.id === userId) {
-      setAdminUsersDeleteError(m.account_admin_users_delete_self_blocked());
-      return;
-    }
-
-    setDeletingAdminUserId(userId);
-    setAdminUsersDeleteError(null);
-
-    try {
-      const response = await fetchFn(`/api/users/${userId}`, {
-        method: "DELETE",
-      });
-      if (!response.ok) {
-        throw new Error((await readErrorDetail(response)) ?? m.account_admin_users_delete_failed());
-      }
-
-      setAdminUsers((current) => current.filter((user) => user.id !== userId));
-      showSuccessToast(m.account_admin_users_deleted(), "bi-trash");
-    } catch (error) {
-      logger.error("Failed to delete admin-managed user:", error);
-      setAdminUsersDeleteError(
-        error instanceof Error && error.message.trim() !== ""
-          ? error.message
-          : m.account_admin_users_delete_failed(),
-      );
-    } finally {
-      setDeletingAdminUserId(null);
-    }
-  };
-
   const handleDeleteAccount = async () => {
     setIsDeletingAccount(true);
     setDeleteAccountError(null);
@@ -254,12 +158,6 @@ export function useSettingsAccount({
     hasProfileChanges,
     resolvedDisplayName,
     handleSaveProfile,
-    adminUsers,
-    isAdminUsersLoading,
-    adminUsersError,
-    adminUsersDeleteError,
-    deletingAdminUserId,
-    handleDeleteAdminUser,
     isDeletingAccount,
     deleteAccountError,
     handleDeleteAccount,
