+++ frontend/src/router.tsx
@@ -22,15 +22,20 @@ const settingsRoute = createRoute({
   validateSearch: (
     search: Record<string, unknown>,
   ): {
-    section?: "general" | "features" | "timeTracking" | "account" | "sync" | "data" | "about";
+    section?: "general" | "features" | "timeTracking" | "account" | "admin" | "data" | "about";
   } => {
     const section = typeof search.section === "string" ? search.section : undefined;
+    // "sync" was folded into "account" when the two settings sections were combined;
+    // keep old bookmarks/links working by redirecting them there.
+    if (section === "sync") {
+      return { section: "account" };
+    }
     if (
       section === "general" ||
       section === "features" ||
       section === "timeTracking" ||
       section === "account" ||
-      section === "sync" ||
+      section === "admin" ||
       section === "data" ||
       section === "about"
     ) {
